remove needless classes
