+++ wheat-modern/src/main/scala/x7c1/wheat/modern/database/presets/presets.scala
@@ -1,7 +1,7 @@
 package x7c1.wheat.modern.database.presets
 
 import android.database.SQLException
-import x7c1.wheat.modern.database.{Findable2, ReadableDatabase, RecordIdentifiable, SeqSelectable2}
+import x7c1.wheat.modern.database.{RecordFindable, ReadableDatabase, RecordIdentifiable, SeqSelectable}
 import x7c1.wheat.modern.either.OptionEither
 import x7c1.wheat.modern.either.Imports._
 
@@ -10,7 +10,7 @@ import scala.language.higherKinds
 trait CollectFrom [I[T] <: RecordIdentifiable[T], A]{
   protected def readable: ReadableDatabase
 
-  def collectFrom[X: I](target: X)(implicit i: SeqSelectable2[I, A]): Either[SQLException, Seq[A]] = {
+  def collectFrom[X: I](target: X)(implicit i: SeqSelectable[I, A]): Either[SQLException, Seq[A]] = {
     readable.select2[Seq[A]] by target
   }
 }
@@ -18,7 +18,7 @@ trait CollectFrom [I[T] <: RecordIdentifiable[T], A]{
 trait Find[I[T] <: RecordIdentifiable[T], A]{
   protected def readable: ReadableDatabase
 
-  def find[X: I](target: X)(implicit i: Findable2[I, A]): OptionEither[SQLException, A] = {
+  def find[X: I](target: X)(implicit i: RecordFindable[I, A]): OptionEither[SQLException, A] = {
     val either = readable.select2[Option[A]] by target
     either.toOptionEither
   }
